dev-util/nvidia-cuda-toolkit-12.9.1-r1 and 13.0.0
Added a new USE flag: "cuda-minor-compat", enabled by default - that
enforces a minimum/maximum nvidia-drivers version as a runtime dependency, as
suggested in: https://bugs.gentoo.org/854345
and as documented in: https://docs.nvidia.com/cuda/cuda-toolkit-release-notes/index.html#cuda-driver

nvidia-cuda-toolkit-13.0.0 depends on a beta driver not packaged yet:
the 580.x series. When it appears, we are ready for it.